requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.0.4
google-auth-oauthlib>=0.4.1
openai
langchain
gradio
python-dotenv